backend/requirements/base.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.1
pydantic-settings==2.8.1
python-jose[cryptography]==3.3.0
httpx==0.28.1
openai==1.72.0
azure-identity==1.21.0
azure-search-documents==11.5.2
mssql-python==0.13.0
jsonschema==4.23.0
python-dotenv==1.1.0
